FIR: refactor annotation-related utils
